More selenium tests.
