--- src/test/gwt/material/design/demo/selenium/test/ui/TabsTest.java
@@ -1,75 +1,0 @@
-/*
- * #%L
- * GwtMaterial
- * %%
- * Copyright (C) 2015 - 2017 GwtMaterialDesign
- * %%
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * 
- *      http://www.apache.org/licenses/LICENSE-2.0
- * 
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- * #L%
- */
-package gwt.material.design.demo.selenium.test.ui;
-
-import gwt.material.design.demo.selenium.test.base.BaseSeleniumTest;
-import gwt.material.design.demo.selenium.test.base.WebDriverManager;
-import org.openqa.selenium.By;
-import org.openqa.selenium.WebElement;
-import org.testng.Assert;
-
-public class TabsTest extends BaseSeleniumTest {
-
-
-    public TabsTest(WebDriverManager driver) {
-        super(driver, "Tabs");
-    }
-
-    @Override
-    protected void runTests() {
-        checkBasic();
-        checkDisabledTabItem();
-        checkSetTabIndex();
-        checkGetTabIndex();
-        checkSelection();
-    }
-
-    protected void checkDisabledTabItem() {
-        scrollToSection(2);
-    }
-
-    protected void checkSetTabIndex() {
-        scrollToSection(4);
-    }
-
-    protected void checkGetTabIndex() {
-        scrollToSection(5);
-    }
-
-    protected void checkSelection() {
-        scrollToSection(9);
-    }
-
-    protected void checkBasic() {
-
-    }
-
-    protected boolean isActive(WebElement element) {
-        return element.findElement(By.cssSelector("a")).getAttribute("class").contains("active");
-    }
-
-    protected WebElement findTab(int index) {
-        return findElement(".code:nth-child("+ index + ")");
-    }
-
-    protected WebElement findTabItem(int tabIndex, int itemIndex) {
-        return findTab(tabIndex).findElement(By.cssSelector("ul.tabs li:nth-child(" + itemIndex + ")"));
-    }
-}
